% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_Context.R
\name{Detect_Context}
\alias{Detect_Context}
\title{Detect sequence context (stranded)}
\usage{
Detect_Context(ref_fasta, chr, pos, context)
}
\arguments{
\item{ref_fasta}{fasta file}

\item{chr}{chromosome(s) to be looked up}

\item{pos}{position(s) to be looked up}

\item{context}{context to be detected, accepts IUPAC nucleotide code,
if more than 1 C in the context, the target C must be denoted by using an X,
for example, "TCX" looks for cytosines preceded by TC. And "XC" looks for cytosines followed by a C.}
}
\description{
This function detects contexts and returns a vector of -1, 1 and 0 showing the presense and absense of the selected context. -1 and +1 shows the strand that cytosine is on.
This function is not optimized for speed.
}
\note{
Corrently this function would return an error for out of bound ranges,
such as positions at the edges of the sequence.
}
\examples{
Detect_Context("./ref.fa",chr="chr2",pos=c(1234,2345),"WRCY") -> out

}
