% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_all.R
\name{duplicated_all}
\alias{duplicated_all}
\title{duplicated_all}
\usage{
duplicated_all(x)
}
\description{
gets all duplicated values
for example in duplicated_all(c("a","b","b","c")), the output will be c(F,T,T,F)
}
