% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loop_seq.R
\name{get_loop_seq}
\alias{get_loop_seq}
\title{Get hairpin loop sequence}
\usage{
get_loop_seq(X)
}
\arguments{
\item{X}{input must be a data.frame of Hairpin_survey output with columns looplen,looppos, minus2,minus1,minus0,plus1,plus2,plus3
output will be of character strings like "(NATC.C.C)" with the targeted C shown as .C.}
}
\description{
gets the sequence of hairpin loops from a data.frame
}
