% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_context_matrix.R
\name{get_mut_context_matrix}
\alias{get_mut_context_matrix}
\title{Get Mutation Context Matrix}
\usage{
get_mut_context_matrix(input, context = "NNCN", UI = FALSE, custom_name = NULL)
}
\arguments{
\item{input}{GRange object conatining the mutations}

\item{context}{defines the extent for with the context of a mutation at C:G position is determined}

\item{UI}{if set to TRUE, the UI values will be calculated instead of counting the mutations}

\item{custom_name}{for when there is only one sample, you can set the name to be displayed}
}
\description{
from Grange objects that lists all the mutations.
This is a wrapper function to use the modified mut_matrix_UI function
This function also allows for assymetrical context to be generated
}
\references{
https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-018-0539-0
}
