% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapacross.R
\name{mapacross}
\alias{mapacross}
\title{Mapacross, similar to Mike Lawrence's Matlab function}
\usage{
mapacross(x, A, B, m = NA)
}
\arguments{
\item{x}{input vector}

\item{A}{the vector to map x to}

\item{B}{the vector the extract ouput from
output will be of from B or m=NA if missing}
}
\description{
map an element to a vector A and return matching value from another vector B
}
