% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_context_matrix.R
\name{mut_96_UI}
\alias{mut_96_UI}
\title{Count mutation occurrences in the provided contexts}
\usage{
mut_96_UI(type_context, gr_sizes, UI = FALSE)
}
\arguments{
\item{type_context}{result from type_context function}

\item{gr_sizes}{A vector indicating the number of variants per GRanges}

\item{UI}{if TRUE, it will calculated the mean UI instead of counting the mutations}
}
\value{
Mutation matrix with 96 trinucleotide mutation occurrences
}
\description{
@details
This function is called by mut_matrix_UI. It is a modification of the mut_96_occurrences() function from the MutationalPatterns package.
It calculates the context for all variants and then splits these per GRanges (samples). It then calculates how often each context occurs.
This function is compatible with assymetric context and UI value
}
\references{
https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-018-0539-0
}
