% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_context_matrix.R
\name{mut_matrix_UI}
\alias{mut_matrix_UI}
\title{Make mutation count matrix}
\usage{
mut_matrix_UI(vcf_list, ref_genome, extension = 1, UI = FALSE)
}
\arguments{
\item{vcf_list}{GRangesList or GRanges object.}

\item{ref_genome}{BSgenome reference genome object}

\item{extension}{The number of bases, that's extracted upstream and
downstream of the base substitutions. (Default: 1).}

\item{UI}{if set to TRUE the mean UI value will be calculated}
}
\value{
mutation count matrix
}
\description{
Make mutation count matrix
this is a modification of the mut_matrix() function from MutationPatterns package
}
\references{
https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-018-0539-0
}
