% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDC.R
\name{NDC}
\alias{NDC}
\alias{ndc}
\title{Normalized Differntial Coverage}
\usage{
NDC(x, y, w = 1e+05, b = 10000, mavn = 120, NDC_Version = 2)
}
\arguments{
\item{x}{coverage vector from the treatment sample}

\item{y}{coverage vector from the control sample}

\item{w}{width of normalization window (for NDC2|3); w=1e5}

\item{b}{step size of w; b =1e4}

\item{mavn}{size of moving average window; mavn = 120}

\item{NDC_Version}{is the version of NDC; NDC_Version = 2}
}
\description{
NDC1 is the original version that uses the whole genome for normalization ; NDC2 divide the genome into smaller parts for normalization ; NDC3 is similar to NDC2 but better for when coverage is 0 at many positions.

default values are optimized based on e.coli experiments
}
\examples{
NDC(treatment$cov , control$cov , w=5e5 , b = 2e4 , mavn = 150 ) -> NDC
}
\keyword{NDC}
\keyword{NDC1}
\keyword{NDC2}
\keyword{NDC3}
\keyword{Normalized}
\keyword{coverage}
\keyword{differential}
