% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiverPlot.R
\name{plot_river2}
\alias{plot_river2}
\title{plot_river2}
\usage{
plot_river2(
  mut_matrix,
  condensed = FALSE,
  UI = FALSE,
  is_normalized = FALSE,
  fontsize = 3,
  alt_colors = FALSE,
  FacetNames = NULL
)
}
\arguments{
\item{mut_matrix}{Matrix containing mutation counts.}

\item{condensed}{More condensed plotting format. Default = F.}

\item{UI}{set to TRUE if ploting Uracilation Index Values}

\item{is_normalized}{set to TRUE if the input data is normalized, to change labels}

\item{fontsize}{font size of the text of the nucleotide/mutation labels}

\item{alt_colors}{if set tot TRUE, alternative colors will be used}

\item{FacetNames}{a character vector to be used for facet labels}

\item{name}{description}
}
\value{
A ggplot object
}
\description{
plot_river2 is modified version of plot_river from MutationalPatterns
}
\examples{

## See the 'mut_matrix()' examples for how we obtained the
## mutation matrix information:
## Get regular matrix
mut_mat <- readRDS(system.file("states/mut_mat_data.rds",
  package = "MutationalPatterns"
))

## Create heatmap of profile
plot_river(mut_mat[,c(1,4)])

## Get extended matrix
mut_mat_extended <- readRDS(system.file("states/mut_mat_data_extended.rds",
  package = "MutationalPatterns"
))

## Create heatmap of extended profile
plot_river(mut_mat_extended[,c(1,4)])

## Create condensed version of riverplot
plot_river(mut_mat_extended[,c(1,4)], condensed = TRUE)

}
\references{
https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-018-0539-0

#' Plot a riverplot

Function to plot a SNV mutation matrix as a riverplot.
This is especially useful when looking at a wide
mutational context
}
\seealso{
\code{\link{mut_matrix}},
\code{\link{plot_96_profile}},
\code{\link{plot_profile_heatmap}}
}
