% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.readcount.R
\name{read.readcount}
\alias{read.readcount}
\title{read readcount output with correct column names}
\usage{
read.readcount(X)
}
\arguments{
\item{Input}{char. path to RC table}
}
\description{
use the folloing line to genereate readcount table:
$ bam-readcount -w0 -d 1000000000 -f ref.fa alignment.sorted.bam | awk -F ":|\t|=" 'BEGIN {OFS = "\t"}; {print $1, $2, $3 , $4, $21 , $35, $49 , $63}' > BRC.txt
}
\examples{
read.readcount("./path/to/RC.txt") -> RC

}
\keyword{BRC}
\keyword{RC}
\keyword{readcount}
