% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiverPlot.R
\name{RivPlot}
\alias{RivPlot}
\title{River Plots}
\usage{
RivPlot(
  input,
  fontsize = 4,
  is_ui = FALSE,
  is_norm = FALSE,
  title = "",
  AltCol = T,
  facet = NULL
)
}
\arguments{
\item{input}{mutation count matrix from the get_mut_context_matrix() function}

\item{fontsize}{font size of the text of the nucleotide/mutation labels}

\item{is_ui}{are you plotting UI values or mutation count?}

\item{is_norm}{is the data normalized?}

\item{title}{Plot's title}

\item{AltCol}{using alternative colors?}

\item{facet}{a character vector to be used for facet labels}
}
\description{
RivPlot is a wrapper for the modified plot_river2 function.
the original plot_river is taken from the MutationalPatterns Package
}
