% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2pats.R
\name{tree2pats}
\alias{tree2pats}
\title{Tree to Pattern}
\usage{
tree2pats(infile)
}
\arguments{
\item{infile}{phylogentic tree}
}
\description{
turn edges(branches) of phylogenetic trees into binary patterns of present/absent of terminal nodes
}
