% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_context_matrix.R
\name{type_context}
\alias{type_context}
\title{Retrieve context of base substitution types}
\usage{
type_context(vcf, ref_genome, extension = 1)
}
\arguments{
\item{vcf}{A CollapsedVCF object}

\item{ref_genome}{Reference genome}

\item{extension}{The number of bases, that's extracted upstream and
downstream of the base substitutions. (Default: 1).}
}
\value{
Mutation types and context character vectors in a named list
}
\description{
A function to extract the bases 3' upstream and 5' downstream of the base
substitution types. This function is from the MutationalPatterns Package.
}
\references{
https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-018-0539-0

#' @export
}
