% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UX.R
\name{UX}
\alias{UX}
\alias{ux}
\title{Calculating Uracilation index}
\usage{
UX(Input, context = "NC")
}
\arguments{
\item{Input}{is readcount data.frame}

\item{context}{the nucleotide context of cytosines; "NC" , "TC" , "WRCY"; Defaults to "NC" for all C:G positions.}
}
\description{
input's column names should be:("Chr.","POS","REF","COV","A","C","G","T")
output is a vector with Uracilation fraction values at C:G positions, the length of the vector is the same as the nrow of the input
this is ~300 times faster than UX_TC for getting only CtoT_GtoA vector
}
\examples{
UX(BRC) -> ui
}
\keyword{Fraction}
\keyword{Index}
\keyword{Uracilation}
