% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UX_TC.R
\name{UX_TC}
\alias{UX_TC}
\alias{ux_tc}
\title{Calculating Uracilation index}
\usage{
UX_TC(Y)
}
\arguments{
\item{Input}{is readcount data.frame}
}
\description{
input's column names should be:("Chr.","POS","REF","COV","A","C","G","T")
output is a list:  CtoT_indx , GtoA_indx , U_indx , Number_of_TCs , Number_of_GAs , Number_of_TC_GA ,  Out = UDF
this runs very slow with big inputs
}
\examples{
UX_TC(BRC) -> ui
}
\keyword{Fraction}
\keyword{Index}
\keyword{Uracilation}
