% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UX_v2.R
\name{UX_v3}
\alias{UX_v3}
\alias{UX_V3}
\alias{ux_v3}
\title{UX version3}
\usage{
UX_v3(X, context = "NC")
}
\arguments{
\item{X}{BRC-style data.frame with columns (POS,REF,COV,A,T,C,G)}

\item{context}{context to be detected, corrently only "NC","TC","WRC" and "WRCY" are covered.}
}
\description{
combines UX function with detect_context_stranded, it filters for position with coverage > 0
new function to calculate uracilation index at the selected context.
}
