% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UX_v2.R
\name{UX_v4}
\alias{UX_v4}
\alias{UX_V3}
\alias{ux_v3}
\alias{UX_v3}
\title{UX version4}
\usage{
UX_v4(X, test_vector)
}
\arguments{
\item{X}{BRC-style data.frame with columns (POS,REF,COV,A,T,C,G)}

\item{test_vector}{numeric vector with 1,-1 and 0 values indicating which rows to include, similar to the output of Detect_Context}
}
\description{
combines UX function with detect_context_stranded, it filters for position with coverage > 0
new function to calculate uracilation index at the selected context.
}
