% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XDisplay.R
\name{XDisplay}
\alias{XDisplay}
\title{Set DISPLAY settings for X11 forwarding}
\usage{
XDisplay(version = 10, from_system = F)
}
\arguments{
\item{version}{int (default=10)}
}
\description{
Set DISPLAY settings for X11 forwarding
}
